+++ infra/terraform/main.tf
@@ -158,7 +158,7 @@ resource "aws_s3_bucket_server_side_encryption_configuration" "factory_artifacts
 
   rule {
     apply_server_side_encryption_by_default {
-      sse_algorithm = "aws:kms"
+      sse_algorithm = "AES256"
     }
   }
 }